+++ src/pages/Details/info/Content.jsx
@@ -1,18 +1,18 @@
 import React, { useRef } from "react";
 
 import SimpleSlider from ".././carousel/SimpleSlider";
-import OrderContainer from ".././ordercontainer/OrderContainer";
+import OrderContainer from ".././order/OrderContainer";
 import styled from "@emotion/styled";
 import MenuTabs from "./MenuTabs";
-import LinkMenu from "../LinkMenu";
-import coat1 from '../../images/coat1.jpg';
-import coat2 from '../../images/coat2.jpg';
-import coat3 from '../../images/coat3.jpg';
-import coat4 from '../../images/coat4.jpg';
-import coat5 from '../../images/coat5.jpg';
-import coat6 from '../../images/coat6.jpg';
-import coat7 from '../../images/coat7.jpg';
-import coat8 from '../../images/coat8.jpg';
+import LinkMenu from "../info/LinkMenu";
+import coat1 from '../../../assets/Details/coat1.jpg';
+import coat2 from '../../../assets/Details/coat2.jpg';
+import coat3 from '../../../assets/Details/coat3.jpg';
+import coat4 from '../../../assets/Details/coat4.jpg';
+import coat5 from '../../../assets/Details/coat5.jpg';
+import coat6 from '../../../assets/Details/coat6.jpg';
+import coat7 from '../../../assets/Details/coat7.jpg';
+import coat8 from '../../../assets/Details/coat8.jpg';
 import RelatedProduct from "./RelatedProduct";
 import MaterialTable from "./MaterialTable";
 import SizeTable from "./SizeTable";
