+++ src/pages/Details/info/MenuTabs.jsx
