+++ src/pages/Details/info/SizeTable.jsx
